[update] Separate files and fixed import
